Fix A-Rows overflow

+++ ydb/core/kqp/opt/kqp_query_plan.cpp
@@ -2789,7 +2789,7 @@ private:
                 // top level rows/size have to match stage output
                 if (!operatorRows && stats.contains("OutputRows")) {
                     auto outputRows = stats.at("OutputRows");
-                    int aRows = outputRows.IsMap() ? outputRows.GetMapSafe().at("Sum").GetIntegerSafe() : outputRows.GetIntegerSafe();
+                    i64 aRows = outputRows.IsMap() ? outputRows.GetMapSafe().at("Sum").GetIntegerSafe() : outputRows.GetIntegerSafe();
                     if (fromBroadcast && parentTaskCount && (aRows % parentTaskCount == 0)) {
                         aRows /= parentTaskCount;
                     }
